requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.1.2
aiofiles>=23.2.1


# AI/ML dependencies
openai>=1.3.0

# PDF processing
pdfplumber>=0.10.0
weasyprint>=60.0
pymupdf
spacy
fitz
pypdf>=4.0.0

# Markdown processing
markdown>=3.5.0

# Development dependencies (optional, can be removed for production)
# jupyterlab
# ipykernel
# ipywidgets
# gradio